1.1.0 -> 1.2.0 - Now when download starts thumbnail is automatically downloaded from youtube - If u download a video or a .mp3, it will be applied the thumbnail - Some bug fixing

+++ main.py
@@ -12,9 +12,17 @@ sys.path.append(vendor_dir)
 
 from pytube import YouTube
 from exitstatus import ExitStatus
+import urllib.request
 
 if platform.system() != "Windows":
     import ffmpeg
+
+
+def clear_screen():
+    if platform.system() == "Windows":
+        subprocess.call('cls', shell=True)
+    else:
+        subprocess.call('clear', shell=True)
 
 
 def main():
@@ -49,10 +57,7 @@ def main():
         if choose == '4':
             break
         elif choose == '1' or choose == '2' or choose == '3':
-            if platform.system() == "Windows":
-                subprocess.call('cls', shell=True)
-            else:
-                subprocess.call('clear', shell=True)
+            clear_screen()
             choosing_video(choose, links)
             break
         else:
@@ -66,10 +71,7 @@ def main():
     sys.exit(ExitStatus.success)
 
 
-def download(yt, separate_tracks=False, tag=None, audio_only=False):
-    # It creates the file name
-    file_name = yt.title + '.mp4'
-
+def acceptable_name_for_ffmpeg(file_name):
     # Check if there are some not enable characters and replaces it
     if file_name.find("\"") != -1:
         file_name = file_name.replace('"', '``')
@@ -82,6 +84,27 @@ def download(yt, separate_tracks=False, tag=None, audio_only=False):
 
     if file_name.find(':') != -1:
         file_name = file_name.replace(':', ' ')
+    return file_name
+
+
+def download(yt, separate_tracks=False, tag=None, audio_only=False):
+    # It creates the file name
+    file_name = yt.title + '.mp4'
+
+    file_name = acceptable_name_for_ffmpeg(file_name)
+
+    # Taking link of thumbnails from youtube
+    thumb_url = yt.thumbnail_url
+
+    # Creates opener for urllib
+    opener = urllib.request.build_opener()
+    opener.addheaders = [('User-Agent', 'Mozilla/5.0 (Windows NT 6.1; WOW64) AppleWebKit/537.36 '
+                                        '(KHTML, like Gecko) Chrome/36.0.1941.0 Safari/537.36')]
+
+    # Install the opener for Urllib request
+    urllib.request.install_opener(opener)
+    # Download image from link
+    urllib.request.urlretrieve(thumb_url, 'thumb.jpg')
 
     # If user choose to select the quality
     if separate_tracks:
@@ -124,8 +147,6 @@ def download(yt, separate_tracks=False, tag=None, audio_only=False):
 
         print("\nMuxing in corso...", end='')
 
-        # Cross-platform ffmpeg
-
         # Create list of file and dir in root
         os.chdir('./')
         listdir = os.listdir(os.getcwd())
@@ -153,11 +174,23 @@ def download(yt, separate_tracks=False, tag=None, audio_only=False):
         # Calls to FFmpeg
         if platform.system() == 'Windows':
             if platform.architecture()[0] == '64bit':
-                command = "ffmpeg_x64 -i {video} -i {audio} -loglevel warning -c:v copy -c:a aac {output}". \
-                    format(video='../' + input_name, audio='../audio_file.mp4', output="\"../" + file_name + "\"")
+                command = "ffmpeg_x64 -i {video} -i {audio}" \
+                          " -loglevel warning" \
+                          " -c:v copy" \
+                          " -c:a aac" \
+                          " {output}" \
+                    .format(video='../' + input_name,
+                            audio='../audio_file.mp4',
+                            output="\"../" + file_name + "\"")
             else:
-                command = "ffmpeg_x86 -i {video} -i {audio} -loglevel warning -c:v copy -c:a aac {output}". \
-                    format(video='../' + input_name, audio='../audio_file.mp4', output="\"../" + file_name + "\"")
+                command = "ffmpeg_x86 -i {video} -i {audio}" \
+                          " -loglevel warning" \
+                          " -c:v copy" \
+                          " -c:a aac" \
+                          " {output}" \
+                    .format(video='../' + input_name,
+                            audio='../audio_file.mp4',
+                            output="\"../" + file_name + "\"")
 
             try:
                 subprocess.call(command, shell=True, cwd='./ffmpeg/')
@@ -174,14 +207,51 @@ def download(yt, separate_tracks=False, tag=None, audio_only=False):
                 print(e.stderr.decode(), file=sys.stderr)
                 alright = False
 
+        os.remove('./audio_file.mp4')
+        os.remove('./' + 'video_file' + extension)
+
         # Check if something went wrong
         if alright:
             print('finito!')
         else:
             print("\nC'è un problema con il muxing del video")
+            os.remove('./' + 'video_file' + extension)
+            return
 
-        os.remove('./' + 'video_file' + extension)
-        os.remove('./audio_file.mp4')
+        print('Creazione anteprima...', end='')
+
+        # Apply thumbnail on video file
+        if platform.system() == 'Windows':
+            if platform.architecture()[0] == '64bit':
+                command = "ffmpeg_x64 -i {video} -i {image}" \
+                          " -loglevel quiet" \
+                          " -map 1 -map 0" \
+                          " -c copy -disposition:0 " \
+                          "attached_pic ../out.mp4"\
+                    .format(video="\"../" + input_name + "\"", image='../thumb.jpg')
+            else:
+                command = "ffmpeg_x86 -i {video} -i {image}" \
+                          " -loglevel quiet" \
+                          " -map 1 -map 0" \
+                          " -c copy -disposition:0 " \
+                          "attached_pic ../out.mp4"\
+                    .format(video="\"../" + input_name + "\"", image='../thumb.jpg')
+
+            try:
+                subprocess.call(command, shell=True, cwd='./ffmpeg/')
+            except Exception as e:
+                print(str(e))
+                alright = False
+
+        if alright:
+            print('finito!')
+            os.remove(input_name)
+            os.rename('out.mp4', input_name)
+        else:
+            print("C'è un problema con la creazione dell'anteprima")
+
+        os.remove('thumb.jpg')
+
     else:
         alright = True
 
@@ -221,11 +291,21 @@ def download(yt, separate_tracks=False, tag=None, audio_only=False):
             # Calls to FFmpeg
             if platform.system() == 'Windows':
                 if platform.architecture()[0] == '64bit':
-                    command = "ffmpeg_x64 -i {video} -f mp3 -ab 192000 -vn {audio} -loglevel warning". \
-                        format(video="\"../" + file_name + "\"", audio="\"../" + file_name.replace('mp4', 'mp3') + "\"")
+                    command = "ffmpeg_x64 -i {video}" \
+                              " -f mp3" \
+                              " -ab 192000" \
+                              " -vn {audio}" \
+                              " -loglevel warning" \
+                        .format(video="\"../" + file_name + "\"",
+                                audio="\"../" + file_name.replace('mp4', 'mp3') + "\"")
                 else:
-                    command = "ffmpeg_x86 -i {video} -f mp3 -ab 192000 -vn {audio} -loglevel warning". \
-                        format(video="\"../" + file_name + "\"", audio="\"../" + file_name.replace('mp4', 'mp3') + "\"")
+                    command = "ffmpeg_x86 -i {video}" \
+                              " -f mp3" \
+                              " -ab 192000" \
+                              " -vn {audio}" \
+                              " -loglevel warning" \
+                        .format(video="\"../" + file_name + "\"",
+                                audio="\"../" + file_name.replace('mp4', 'mp3') + "\"")
 
                 try:
                     subprocess.call(command, shell=True, cwd='./ffmpeg/')
@@ -252,7 +332,45 @@ def download(yt, separate_tracks=False, tag=None, audio_only=False):
                 print('finito!')
             else:
                 print("\nProblema con la conversione del file")
-            os.remove(file_name)
+
+            # Apply thumbnail on video file
+            print('Creazione anteprima...', end='')
+
+            if platform.system() == 'Windows':
+                if platform.architecture()[0] == '64bit':
+                    command = "ffmpeg_x64 -i {audio} -i {image}" \
+                              " -loglevel quiet" \
+                              " -map 0:0 -map 1:0" \
+                              " -c:v copy -c:a copy " \
+                              "-id3v2_version 3 " \
+                              "../out.mp3"\
+                        .format(audio="\"../" + file_name.replace('mp4', 'mp3') + "\"", image='../thumb.jpg')
+                else:
+                    command = "ffmpeg_x86 -i {audio} -i {image}" \
+                              " -loglevel quiet" \
+                              " -map 0:0 -map 1:0" \
+                              " -c:v copy -c:a copy " \
+                              "-id3v2_version 3 " \
+                              "../out.mp3"\
+                        .format(audio="\"../" + file_name.replace('mp4', 'mp3') + "\"", image='../thumb.jpg')
+
+                try:
+                    subprocess.call(command, shell=True, cwd='./ffmpeg/')
+                except Exception as e:
+                    print(str(e))
+                    alright = False
+
+            if alright:
+                print('finito!')
+                os.remove(file_name)
+                file_name = file_name.replace('mp4', 'mp3')
+                os.remove(file_name)
+                os.rename('out.mp3', file_name)
+            else:
+                print("C'è un problema con la creazione dell'anteprima")
+
+            os.remove('thumb.jpg')
+
         else:
             # If user wants to download in 720p
             try:
@@ -266,11 +384,55 @@ def download(yt, separate_tracks=False, tag=None, audio_only=False):
                 except Exception as e:
                     print(str(e))
                     alright = False
+                else:
+                    print('finito!')
+                    file_name = yt.title + '.mp4'
+                    file_name = acceptable_name_for_ffmpeg(file_name)
+
+                    # Create list of file and dir in root
+                    os.chdir('./')
+                    listdir = os.listdir(os.getcwd())
+
+                    for i in listdir:
+                        if i.find('.mp4') != -1:
+                            os.rename(i, file_name)
+
+                    # Apply thumbnail on video file
+                    print('Creazione anteprima...', end='')
+
+                    if platform.system() == 'Windows':
+                        if platform.architecture()[0] == '64bit':
+                            command = "ffmpeg_x64 -i {video} -i {image}" \
+                                      "-loglevel warning " \
+                                      "-map 1 -map 0 " \
+                                      "-c copy -disposition:0 " \
+                                      "attached_pic ../out.mp4"\
+                                .format(video="\"../" + file_name + "\"", image='../thumb.jpg')
+                        else:
+                            command = "ffmpeg_x64 -i {video} -i {image} " \
+                                      "-loglevel warning " \
+                                      "-map 1 -map 0 " \
+                                      "-c copy -disposition:0 " \
+                                      "attached_pic ../out.mp4"\
+                                .format(video="\"../" + file_name + "\"", image='../thumb.jpg')
+
+                        try:
+                            subprocess.call(command, shell=True, cwd='./ffmpeg/')
+                        except Exception as e:
+                            print(str(e))
+                            alright = False
+
+                    if alright:
+                        print('finito!')
+                        os.remove(file_name)
+                        os.rename('out.mp4', file_name)
+                    else:
+                        print("C'è un problema con la creazione dell'anteprima")
+
+                    os.remove('thumb.jpg')
 
             # Check is something went wrong
-            if alright:
-                print('finito!')
-            else:
+            if not alright:
                 print("\nProblema con il download del file")
 
 
@@ -290,10 +452,12 @@ def choosing_video(choose, links):
 
         # Taking info about video
         try:
+            print('Caricamento in corso...')
             yt = YouTube(link)
         except Exception as e:
             print_error(str(e))
         else:
+            clear_screen()
             print("\nDettagli del video ", i + 1, "\n")
             print("Titolo del video:          ", yt.title)
             print("Numero di visualizzazioni: ", yt.views)
